+++ package.json
@@ -72,7 +72,9 @@
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
     "@babel/plugin-proposal-optional-chaining": "^7.21.0",
     "@babel/preset-env": "^7.21.4",
+    "@vue/cli-plugin-e2e-webdriverio": "^5.0.8",
     "@vue/cli-plugin-router": "~4.4.0",
+    "@vue/cli-plugin-unit-jest": "^5.0.8",
     "@vue/cli-service": "~4.4.0",
     "@vue/devtools": "^6.5.0",
     "@vue/eslint-config-standard": "^5.1.2",
@@ -87,6 +89,7 @@
     "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-standard": "^4.0.0",
     "eslint-plugin-vue": "^7.8.0",
+    "jest": "^29.5.0",
     "node-sass": "^7.0.1",
     "sass-loader": "^10.2.1",
     "vue-cli-plugin-electron-builder": "~2.1.1",
